[#403] refactor: class, package 이름 변경
