+++ src/main/java/com/modeunsa/boundedcontext/payment/app/usecase/PaymentCompleteUseCase.java
@@ -1,7 +1,7 @@
 package com.modeunsa.boundedcontext.payment.app.usecase;
 
 import com.modeunsa.boundedcontext.payment.app.dto.PaymentProcessContext;
-import com.modeunsa.boundedcontext.payment.app.usecase.settlement.PaymentSettlementRegistry;
+import com.modeunsa.boundedcontext.payment.app.usecase.complete.PaymentCompleteRegistry;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -9,11 +9,11 @@ import org.springframework.transaction.annotation.Transactional;
 @Service
 @Transactional
 @RequiredArgsConstructor
-public class PaymentSettlementUseCase {
+public class PaymentCompleteUseCase {
 
-  private final PaymentSettlementRegistry paymentSettlementRegistry;
+  private final PaymentCompleteRegistry paymentCompleteRegistry;
 
   public void execute(PaymentProcessContext context) {
-    paymentSettlementRegistry.get(context.paymentPurpose()).execute(context);
+    paymentCompleteRegistry.get(context.paymentPurpose()).execute(context);
   }
 }
