+++ src/main/java/com/modeunsa/boundedcontext/payment/app/usecase/complete/PaymentCompleteProcess.java
@@ -1,9 +1,9 @@
-package com.modeunsa.boundedcontext.payment.app.usecase.settlement;
+package com.modeunsa.boundedcontext.payment.app.usecase.complete;
 
 import com.modeunsa.boundedcontext.payment.app.dto.PaymentProcessContext;
 import com.modeunsa.boundedcontext.payment.domain.types.PaymentPurpose;
 
-public interface PaymentSettlementProcess {
+public interface PaymentCompleteProcess {
   PaymentPurpose purpose();
 
   void execute(PaymentProcessContext context);
